Move signal processing functions and sigaction calls to signals from redshift.c
